improved consignments

+++ app/Http/Models/Consignment.php
@@ -48,7 +48,7 @@ class Consignment extends Model
     //PERSONALIZED FUNCTIONS
     public function set_agreement($file)
     {
-        if($this->agreement != '')
+        if($this->agreement && $this->agreement != '')
             Util::remove_file ($this->agreement);
         $this->agreement = Util::upload_file ($file,'consignment');
     }
